Separated ContextOwnerContentProvder and ContextContentsContentProvider
This allows showing only the interesting parts of the context. In some
cases it might be useful to see the context owner (e.g. when having a
context tree only) in other it might not (swing component tree, as the
owner is the component to which it has been attached).